requirements.txt
aiohttp==2.3.7
discord.py==0.16.12